requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
langchain
langchain-google-genai
google-generativeai
pytesseract
python-docx
pypdf
Pillow
requests
pandas
numpy